+++ src/test/java/roomescape/JdbcReservationDaoTest.java
@@ -17,7 +17,7 @@ import roomescape.dto.ReservationRequest;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class ReservationDaoTest {
+class JdbcReservationDaoTest {
 
     @Autowired
     private JdbcTemplate jdbcTemplate;
